refactor predicate to different package
