+++ pkg/operator/controller/garden/care/add_test.go
@@ -14,8 +14,6 @@ import (
 	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	fakeclient "sigs.k8s.io/controller-runtime/pkg/client/fake"
-	"sigs.k8s.io/controller-runtime/pkg/event"
-	"sigs.k8s.io/controller-runtime/pkg/predicate"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 
 	gardencorev1beta1 "github.com/gardener/gardener/pkg/apis/core/v1beta1"
@@ -47,92 +45,6 @@ var _ = Describe("Add", func() {
 		reconciler = &Reconciler{
 			RuntimeClient: runtimeClient,
 		}
-	})
-
-	Describe("#GardenPredicate", func() {
-		var p predicate.Predicate
-
-		BeforeEach(func() {
-			p = reconciler.GardenPredicate()
-		})
-
-		Describe("#Create", func() {
-			It("should return true", func() {
-				Expect(p.Create(event.CreateEvent{})).To(BeTrue())
-			})
-		})
-
-		Describe("#Update", func() {
-			It("should return false because new object is no garden", func() {
-				Expect(p.Update(event.UpdateEvent{})).To(BeFalse())
-			})
-
-			It("should return false because old object is no garden", func() {
-				Expect(p.Update(event.UpdateEvent{ObjectNew: garden})).To(BeFalse())
-			})
-
-			It("should return false because last operation is nil on old garden", func() {
-				Expect(p.Update(event.UpdateEvent{ObjectOld: garden, ObjectNew: garden})).To(BeFalse())
-			})
-
-			It("should return false because last operation is nil on new garden", func() {
-				oldShoot := garden.DeepCopy()
-				oldShoot.Status.LastOperation = &gardencorev1beta1.LastOperation{}
-				Expect(p.Update(event.UpdateEvent{ObjectOld: oldShoot, ObjectNew: garden})).To(BeFalse())
-			})
-
-			It("should return false because last operation type is 'Delete' on old garden", func() {
-				garden.Status.LastOperation = &gardencorev1beta1.LastOperation{}
-				oldShoot := garden.DeepCopy()
-				oldShoot.Status.LastOperation.Type = gardencorev1beta1.LastOperationTypeDelete
-				Expect(p.Update(event.UpdateEvent{ObjectOld: oldShoot, ObjectNew: garden})).To(BeFalse())
-			})
-
-			It("should return false because last operation type is 'Delete' on new garden", func() {
-				garden.Status.LastOperation = &gardencorev1beta1.LastOperation{}
-				garden.Status.LastOperation.Type = gardencorev1beta1.LastOperationTypeDelete
-				oldShoot := garden.DeepCopy()
-				Expect(p.Update(event.UpdateEvent{ObjectOld: oldShoot, ObjectNew: garden})).To(BeFalse())
-			})
-
-			It("should return false because last operation type is not 'Processing' on old garden", func() {
-				garden.Status.LastOperation = &gardencorev1beta1.LastOperation{}
-				garden.Status.LastOperation.Type = gardencorev1beta1.LastOperationTypeReconcile
-				garden.Status.LastOperation.State = gardencorev1beta1.LastOperationStateSucceeded
-				oldShoot := garden.DeepCopy()
-				Expect(p.Update(event.UpdateEvent{ObjectOld: oldShoot, ObjectNew: garden})).To(BeFalse())
-			})
-
-			It("should return false because last operation type is not 'Succeeded' on new garden", func() {
-				garden.Status.LastOperation = &gardencorev1beta1.LastOperation{}
-				garden.Status.LastOperation.Type = gardencorev1beta1.LastOperationTypeReconcile
-				garden.Status.LastOperation.State = gardencorev1beta1.LastOperationStateProcessing
-				oldShoot := garden.DeepCopy()
-				oldShoot.Status.LastOperation.State = gardencorev1beta1.LastOperationStateProcessing
-				Expect(p.Update(event.UpdateEvent{ObjectOld: oldShoot, ObjectNew: garden})).To(BeFalse())
-			})
-
-			It("should return true because last operation type is 'Succeeded' on new garden", func() {
-				garden.Status.LastOperation = &gardencorev1beta1.LastOperation{}
-				garden.Status.LastOperation.Type = gardencorev1beta1.LastOperationTypeReconcile
-				garden.Status.LastOperation.State = gardencorev1beta1.LastOperationStateSucceeded
-				oldShoot := garden.DeepCopy()
-				oldShoot.Status.LastOperation.State = gardencorev1beta1.LastOperationStateProcessing
-				Expect(p.Update(event.UpdateEvent{ObjectOld: oldShoot, ObjectNew: garden})).To(BeTrue())
-			})
-		})
-
-		Describe("#Delete", func() {
-			It("should return false", func() {
-				Expect(p.Delete(event.DeleteEvent{})).To(BeFalse())
-			})
-		})
-
-		Describe("#Generic", func() {
-			It("should return false", func() {
-				Expect(p.Generic(event.GenericEvent{})).To(BeFalse())
-			})
-		})
 	})
 
 	Describe("#MapManagedResourceToGarden", func() {
